requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
python-dotenv
streamlit
requests
